# go-gost/x/api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    APIConfig:
        properties:
            accesslog:
                type: boolean
                x-go-name: AccessLog
            addr:
                type: string
                x-go-name: Addr
            auth:
                $ref: '#/definitions/AuthConfig'
            auther:
                type: string
                x-go-name: Auther
            pathPrefix:
                type: string
                x-go-name: PathPrefix
        type: object
        x-go-package: github.com/go-gost/x/config
    AdmissionConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            matchers:
                items:
                    type: string
                type: array
                x-go-name: Matchers
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
            reverse:
                description: 'Deprecated: use whitelist instead'
                type: boolean
                x-go-name: Reverse
            whitelist:
                type: boolean
                x-go-name: Whitelist
        type: object
        x-go-package: github.com/go-gost/x/config
    AuthConfig:
        properties:
            password:
                type: string
                x-go-name: Password
            username:
                type: string
                x-go-name: Username
        type: object
        x-go-package: github.com/go-gost/x/config
    AutherConfig:
        properties:
            auths:
                items:
                    $ref: '#/definitions/AuthConfig'
                type: array
                x-go-name: Auths
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/go-gost/x/config
    BypassConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            matchers:
                items:
                    type: string
                type: array
                x-go-name: Matchers
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
            reverse:
                description: 'Deprecated: use whitelist instead'
                type: boolean
                x-go-name: Reverse
            whitelist:
                type: boolean
                x-go-name: Whitelist
        type: object
        x-go-package: github.com/go-gost/x/config
    ChainConfig:
        properties:
            hops:
                items:
                    $ref: '#/definitions/HopConfig'
                type: array
                x-go-name: Hops
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: github.com/go-gost/x/config
...
    Config:
        properties:
            admissions:
                items:
                    $ref: '#/definitions/AdmissionConfig'
                type: array
                x-go-name: Admissions
            api:
                $ref: '#/definitions/APIConfig'
            authers:
                items:
                    $ref: '#/definitions/AutherConfig'
                type: array
                x-go-name: Authers
            bypasses:
                items:
                    $ref: '#/definitions/BypassConfig'
                type: array
                x-go-name: Bypasses
            chains:
                items:
                    $ref: '#/definitions/ChainConfig'
                type: array
                x-go-name: Chains
            climiters:
                items:
                    $ref: '#/definitions/LimiterConfig'
                type: array
                x-go-name: CLimiters
            hops:
                items:
                    $ref: '#/definitions/HopConfig'
                type: array
                x-go-name: Hops
            hosts:
                items:
                    $ref: '#/definitions/HostsConfig'
                type: array
                x-go-name: Hosts
            ingresses:
                items:
                    $ref: '#/definitions/IngressConfig'
                type: array
                x-go-name: Ingresses
            limiters:
                items:
                    $ref: '#/definitions/LimiterConfig'
                type: array
                x-go-name: Limiters
            log:
                $ref: '#/definitions/LogConfig'
            loggers:
                items:
                    $ref: '#/definitions/LoggerConfig'
                type: array
                x-go-name: Loggers
            metrics:
                $ref: '#/definitions/MetricsConfig'
            observers:
                items:
                    $ref: '#/definitions/ObserverConfig'
                type: array
                x-go-name: Observers
            profiling:
                $ref: '#/definitions/ProfilingConfig'
            recorders:
                items:
                    $ref: '#/definitions/RecorderConfig'
                type: array
                x-go-name: Recorders
            resolvers:
                items:
                    $ref: '#/definitions/ResolverConfig'
                type: array
                x-go-name: Resolvers
            rlimiters:
                items:
                    $ref: '#/definitions/LimiterConfig'
                type: array
                x-go-name: RLimiters
            routers:
                items:
                    $ref: '#/definitions/RouterConfig'
                type: array
                x-go-name: Routers
            sds:
                items:
                    $ref: '#/definitions/SDConfig'
                type: array
                x-go-name: SDs
            services:
                items:
                    $ref: '#/definitions/ServiceConfig'
                type: array
                x-go-name: Services
            tls:
                $ref: '#/definitions/TLSConfig'
        type: object
        x-go-package: github.com/go-gost/x/config
...
    Duration:
        description: |-
            A Duration represents the elapsed time between two instants
            as an int64 nanosecond count. The representation limits the
            largest representable duration to approximately 290 years.
        format: int64
        type: integer
        x-go-package: time
    FileLoader:
        properties:
            path:
                type: string
                x-go-name: Path
        type: object
        x-go-package: github.com/go-gost/x/config
    FileRecorder:
        properties:
            path:
                type: string
                x-go-name: Path
            sep:
                type: string
                x-go-name: Sep
        type: object
        x-go-package: github.com/go-gost/x/config
...
    ForwarderConfig:
        properties:
            hop:
                description: the referenced hop name
                type: string
                x-go-name: Hop
            name:
                description: 'Deprecated: use hop instead'
                type: string
                x-go-name: Name
            nodes:
                items:
                    $ref: '#/definitions/ForwardNodeConfig'
                type: array
                x-go-name: Nodes
            selector:
                $ref: '#/definitions/SelectorConfig'
        type: object
        x-go-package: github.com/go-gost/x/config
...
    HTTPLoader:
        properties:
            timeout:
                $ref: '#/definitions/Duration'
            url:
                type: string
                x-go-name: URL
        type: object
        x-go-package: github.com/go-gost/x/config
...
    HTTPRecorder:
        properties:
            header:
                additionalProperties:
                    type: string
                type: object
                x-go-name: Header
            timeout:
                $ref: '#/definitions/Duration'
            url:
                type: string
                x-go-name: URL
        type: object
        x-go-package: github.com/go-gost/x/config
...
    HandlerConfig:
        properties:
            auth:
                $ref: '#/definitions/AuthConfig'
            auther:
                type: string
                x-go-name: Auther
            authers:
                items:
                    type: string
                type: array
                x-go-name: Authers
            chain:
                type: string
                x-go-name: Chain
            chainGroup:
                $ref: '#/definitions/ChainGroupConfig'
            limiter:
                type: string
                x-go-name: Limiter
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            observer:
                type: string
                x-go-name: Observer
            retries:
                format: int64
                type: integer
                x-go-name: Retries
            tls:
                $ref: '#/definitions/TLSConfig'
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/go-gost/x/config
    HopConfig:
        properties:
            bypass:
                type: string
                x-go-name: Bypass
            bypasses:
                items:
                    type: string
                type: array
                x-go-name: Bypasses
            file:
                $ref: '#/definitions/FileLoader'
            hosts:
                type: string
                x-go-name: Hosts
            http:
                $ref: '#/definitions/HTTPLoader'
            interface:
                type: string
                x-go-name: Interface
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
            nodes:
                items:
                    $ref: '#/definitions/NodeConfig'
                type: array
                x-go-name: Nodes
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
            resolver:
                type: string
                x-go-name: Resolver
            selector:
                $ref: '#/definitions/SelectorConfig'
            sockopts:
                $ref: '#/definitions/SockOptsConfig'
        type: object
        x-go-package: github.com/go-gost/x/config
    HostMappingConfig:
        properties:
            aliases:
                items:
                    type: string
                type: array
                x-go-name: Aliases
            hostname:
                type: string
                x-go-name: Hostname
            ip:
                type: string
                x-go-name: IP
        type: object
        x-go-package: github.com/go-gost/x/config
    HostsConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            mappings:
                items:
                    $ref: '#/definitions/HostMappingConfig'
                type: array
                x-go-name: Mappings
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/go-gost/x/config
    IngressConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
            rules:
                items:
                    $ref: '#/definitions/IngressRuleConfig'
                type: array
                x-go-name: Rules
        type: object
        x-go-package: github.com/go-gost/x/config
    IngressRuleConfig:
        properties:
            endpoint:
                type: string
                x-go-name: Endpoint
            hostname:
                type: string
                x-go-name: Hostname
        type: object
        x-go-package: github.com/go-gost/x/config
    LimiterConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            limits:
                items:
                    type: string
                type: array
                x-go-name: Limits
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/go-gost/x/config
    ListenerConfig:
        properties:
            auth:
                $ref: '#/definitions/AuthConfig'
            auther:
                type: string
                x-go-name: Auther
            authers:
                items:
                    type: string
                type: array
                x-go-name: Authers
            chain:
                type: string
                x-go-name: Chain
            chainGroup:
                $ref: '#/definitions/ChainGroupConfig'
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            tls:
                $ref: '#/definitions/TLSConfig'
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/go-gost/x/config
    LogConfig:
        properties:
            format:
                type: string
                x-go-name: Format
            level:
                type: string
                x-go-name: Level
            output:
                type: string
                x-go-name: Output
            rotation:
                $ref: '#/definitions/LogRotationConfig'
        type: object
        x-go-package: github.com/go-gost/x/config
    LogRotationConfig:
        properties:
            compress:
                description: |-
                    Compress determines if the rotated log files should be compressed
                    using gzip. The default is not to perform compression.
                type: boolean
                x-go-name: Compress
            localTime:
                description: |-
                    LocalTime determines if the time used for formatting the timestamps in
                    backup files is the computer's local time. The default is to use UTC
                    time.
                type: boolean
                x-go-name: LocalTime
            maxAge:
                description: |-
                    MaxAge is the maximum number of days to retain old log files based on the
                    timestamp encoded in their filename.  Note that a day is defined as 24
                    hours and may not exactly correspond to calendar days due to daylight
                    savings, leap seconds, etc. The default is not to remove old log files
                    based on age.
                format: int64
                type: integer
                x-go-name: MaxAge
            maxBackups:
                description: |-
                    MaxBackups is the maximum number of old log files to retain.  The default
                    is to retain all old log files (though MaxAge may still cause them to get
                    deleted.)
                format: int64
                type: integer
                x-go-name: MaxBackups
            maxSize:
                description: |-
                    MaxSize is the maximum size in megabytes of the log file before it gets
                    rotated. It defaults to 100 megabytes.
                format: int64
                type: integer
                x-go-name: MaxSize
        type: object
        x-go-package: github.com/go-gost/x/config
    LoggerConfig:
        properties:
            log:
                $ref: '#/definitions/LogConfig'
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: github.com/go-gost/x/config
    MetricsConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            auth:
                $ref: '#/definitions/AuthConfig'
            auther:
                type: string
                x-go-name: Auther
            path:
                type: string
                x-go-name: Path
        type: object
        x-go-package: github.com/go-gost/x/config
    NameserverConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            async:
                type: boolean
                x-go-name: Async
            chain:
                type: string
                x-go-name: Chain
            clientIP:
                type: string
                x-go-name: ClientIP
            hostname:
                type: string
                x-go-name: Hostname
            only:
                type: string
                x-go-name: Only
            prefer:
                type: string
                x-go-name: Prefer
            timeout:
                $ref: '#/definitions/Duration'
            ttl:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/go-gost/x/config
    NodeConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            bypass:
                type: string
                x-go-name: Bypass
            bypasses:
                items:
                    type: string
                type: array
                x-go-name: Bypasses
            connector:
                $ref: '#/definitions/ConnectorConfig'
            dialer:
                $ref: '#/definitions/DialerConfig'
            filter:
                $ref: '#/definitions/NodeFilterConfig'
            hosts:
                type: string
                x-go-name: Hosts
            http:
                $ref: '#/definitions/HTTPNodeConfig'
            interface:
                type: string
                x-go-name: Interface
            matcher:
                $ref: '#/definitions/NodeMatcherConfig'
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
            netns:
                type: string
                x-go-name: Netns
            network:
                type: string
                x-go-name: Network
            resolver:
                type: string
                x-go-name: Resolver
            sockopts:
                $ref: '#/definitions/SockOptsConfig'
            tls:
                $ref: '#/definitions/TLSNodeConfig'
        type: object
        x-go-package: github.com/go-gost/x/config
...
    ObserverConfig:
        properties:
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
        type: object
        x-go-package: github.com/go-gost/x/config
    PluginConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            timeout:
                $ref: '#/definitions/Duration'
            tls:
                $ref: '#/definitions/TLSConfig'
            token:
                type: string
                x-go-name: Token
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/go-gost/x/config
    ProfilingConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
        type: object
        x-go-package: github.com/go-gost/x/config
    RecorderConfig:
        properties:
            file:
                $ref: '#/definitions/FileRecorder'
            http:
                $ref: '#/definitions/HTTPRecorder'
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisRecorder'
            tcp:
                $ref: '#/definitions/TCPRecorder'
        type: object
        x-go-package: github.com/go-gost/x/config
    RecorderObject:
        properties:
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
            record:
                type: string
                x-go-name: Record
        type: object
        x-go-package: github.com/go-gost/x/config
    RedisLoader:
        properties:
            addr:
                type: string
                x-go-name: Addr
            db:
                format: int64
                type: integer
                x-go-name: DB
            key:
                type: string
                x-go-name: Key
            password:
                type: string
                x-go-name: Password
            type:
                type: string
                x-go-name: Type
            username:
                type: string
                x-go-name: Username
        type: object
        x-go-package: github.com/go-gost/x/config
    RedisRecorder:
        properties:
            addr:
                type: string
                x-go-name: Addr
            db:
                format: int64
                type: integer
                x-go-name: DB
            key:
                type: string
                x-go-name: Key
            password:
                type: string
                x-go-name: Password
            type:
                type: string
                x-go-name: Type
            username:
                type: string
                x-go-name: Username
        type: object
        x-go-package: github.com/go-gost/x/config
    ResolverConfig:
        properties:
            name:
                type: string
                x-go-name: Name
            nameservers:
                items:
                    $ref: '#/definitions/NameserverConfig'
                type: array
                x-go-name: Nameservers
            plugin:
                $ref: '#/definitions/PluginConfig'
        type: object
        x-go-package: github.com/go-gost/x/config
    Response:
        properties:
            code:
                format: int64
                type: integer
                x-go-name: Code
            msg:
                type: string
                x-go-name: Msg
        type: object
        x-go-package: github.com/go-gost/x/api
    RouterConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
            routes:
                items:
                    $ref: '#/definitions/RouterRouteConfig'
                type: array
                x-go-name: Routes
        type: object
        x-go-package: github.com/go-gost/x/config
    RouterRouteConfig:
        properties:
            gateway:
                type: string
                x-go-name: Gateway
            net:
                type: string
                x-go-name: Net
        type: object
        x-go-package: github.com/go-gost/x/config
    SDConfig:
        properties:
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
        type: object
        x-go-package: github.com/go-gost/x/config
    SelectorConfig:
        properties:
            failTimeout:
                $ref: '#/definitions/Duration'
            maxFails:
                format: int64
                type: integer
                x-go-name: MaxFails
            strategy:
                type: string
                x-go-name: Strategy
        type: object
        x-go-package: github.com/go-gost/x/config
    ServiceConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            admission:
                type: string
                x-go-name: Admission
            admissions:
                items:
                    type: string
                type: array
                x-go-name: Admissions
            bypass:
                type: string
                x-go-name: Bypass
            bypasses:
                items:
                    type: string
                type: array
                x-go-name: Bypasses
            climiter:
                type: string
                x-go-name: CLimiter
            forwarder:
                $ref: '#/definitions/ForwarderConfig'
            handler:
                $ref: '#/definitions/HandlerConfig'
            hosts:
                type: string
                x-go-name: Hosts
            interface:
                description: 'Deprecated: use metadata.interface instead'
                type: string
                x-go-name: Interface
            limiter:
                type: string
                x-go-name: Limiter
            listener:
                $ref: '#/definitions/ListenerConfig'
            logger:
                type: string
                x-go-name: Logger
            loggers:
                items:
                    type: string
                type: array
                x-go-name: Loggers
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
            observer:
                type: string
                x-go-name: Observer
            recorders:
                items:
                    $ref: '#/definitions/RecorderObject'
                type: array
                x-go-name: Recorders
            resolver:
                type: string
                x-go-name: Resolver
            rlimiter:
                type: string
                x-go-name: RLimiter
            sockopts:
                $ref: '#/definitions/SockOptsConfig'
            status:
                $ref: '#/definitions/ServiceStatus'
        type: object
        x-go-package: github.com/go-gost/x/config
...
    ServiceStatus:
        properties:
            createTime:
                format: int64
                type: integer
                x-go-name: CreateTime
            events:
                items:
                    $ref: '#/definitions/ServiceEvent'
                type: array
                x-go-name: Events
            state:
                type: string
                x-go-name: State
            stats:
                $ref: '#/definitions/ServiceStats'
        type: object
        x-go-package: github.com/go-gost/x/config
    SockOptsConfig:
        properties:
            mark:
                format: int64
                type: integer
                x-go-name: Mark
        type: object
        x-go-package: github.com/go-gost/x/config
    TCPRecorder:
        properties:
            addr:
                type: string
                x-go-name: Addr
            timeout:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/go-gost/x/config
    TLSConfig:
        properties:
            caFile:
                type: string
                x-go-name: CAFile
            certFile:
                type: string
                x-go-name: CertFile
            commonName:
                type: string
                x-go-name: CommonName
            keyFile:
                type: string
                x-go-name: KeyFile
            options:
                $ref: '#/definitions/TLSOptions'
            organization:
                type: string
                x-go-name: Organization
            secure:
                type: boolean
                x-go-name: Secure
            serverName:
                type: string
                x-go-name: ServerName
            validity:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/go-gost/x/config
...
info:
    title: Documentation of Web API.
    version: 1.0.0
paths:
    /config:
        get:
            operationId: getConfigRequest
            parameters:
                - description: output format, one of yaml|json, default is json.
                  in: query
                  name: format
                  type: string
                  x-go-name: Format
            responses:
                "200":
                    $ref: '#/responses/getConfigResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Get current config.
            tags:
                - Config
        post:
            operationId: saveConfigRequest
            parameters:
                - description: output format, one of yaml|json, default is yaml.
                  in: query
                  name: format
                  type: string
                  x-go-name: Format
                - description: file path, default is flux.yaml|flux.json in current working directory.
                  in: query
                  name: path
                  type: string
                  x-go-name: Path
            responses:
                "200":
                    $ref: '#/responses/saveConfigResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Save current config to file (flux.yaml or flux.json).
            tags:
                - Config
...
produces:
    - application/json
responses:
...
    getConfigResponse:
        description: successful operation.
        headers:
            Config: {}
        schema:
            $ref: '#/definitions/Config'
...
    saveConfigResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
schemes:
    - https
    - http
securityDefinitions:
    basicAuth:
        type: basic
swagger: "2.0"